fix:condition if the token is there then we make a request when the page is reloaded
