+++ src/components/app/App.js
@@ -1,4 +1,4 @@
-import {useSelector} from "react-redux";
+import {useDispatch, useSelector} from "react-redux";
 import {Routes, Route} from "react-router-dom";
 import {ToastContainer} from 'react-toastify';
 import Dashboard from "../../pages/dashboard/Dashboard";
@@ -6,16 +6,28 @@ import NotFoundPage from "../../pages/notFound";
 import RegisterPage from "../../pages/registerPage";
 import LoginPage from "../../pages/loginPage";
 import LoadingSpinner from "../loadingSpinner";
-import {getIsLoading, getIsLoggedIn} from "../../redux/auth/authSelectors";
+import {getAccessToken, getIsLoading, getIsLoggedIn} from "../../redux/auth/authSelectors";
 import PrivateRoute from "../routes/PrivateRoute";
 import PublicRoute from "../routes/PublicRoute";
 import Header from "../header";
+import {useEffect} from "react";
+import tasksOperations from "../../redux/tasks/tasksOperations";
 
 // import {UserMenu} from "../UserMenu";
 
 function App() {
 	const isLoading = useSelector(getIsLoading);
 	const isLoggedIn = useSelector(getIsLoggedIn);
+	const dispatch = useDispatch();
+	
+	const token = useSelector(getAccessToken);
+	
+	
+	useEffect(() => {
+		if (token) {
+			dispatch(tasksOperations.getTasks());
+		}
+	}, [dispatch, token])
 	
 	return (
 		<>
